use data from hook unconditionally and rename hook accordingly
