+++ packages/dataviews/src/hooks/index.ts
@@ -2,5 +2,5 @@
  * Internal dependencies
  */
 export { default as useFormValidity } from './use-form-validity';
-export { useInfiniteScrollData } from './use-infinite-scroll-data';
+export { useData } from './use-data';
 export { default as useSelectedItems } from './use-selected-items';
